Adding InputValidation for KeyValue secrets
Ran into errors when a storePath contained secrets that were not in the correct format. If the secret did not contain the keys PUBLIC_KEY, and PRIVATE_KEY inventory jobs would stop and discontinue. (At the point it met the "unexpected" secret)

Code implemented here will ignore secrets that do not include the correct keys in the "secret".

Added a note in regards to this to the documentation.